fix: Pass --paging=always to bat (#142)
This gets rid of a slightly annoying difference between files that fit
in the preview pane and ones that don't: hitting the scroll-preview
mappings would complain about a closed stream. Always using the pager
ensures there's always a process at the other end, even if asking it
to scroll doesn't actually do anything.