fix: Ensure git_files uses the git root directory as cwd
This sets the cwd option of the git_files builtin to use the root of the
git directory when the cwd option doesn't already exist. When git lists
files, it is relative to the root of the git directory, rather than the
current working directory. This caused problems when using git_files in
a subdirectory of the git root (see #174).

This commit fixes the issue by always setting the cwd as the root of the
git directory.