Extend previewers for git and differ their mode
This commit includes two groups of changes. First the previewer used for
git commit and bcommit has been extended by two additional ones. While
so far it was only showing the diff to the parent commit, it now also
allows to show the diff to the current buffers state and showing the
file as it was for this commit without any diff.

As second change do all previewers now differ between the commit and
bcommit scenario. For the commit case it shows the diff for all affected
files, while the bcommit case only focuses on the current buffer alone.

Note that the 'commit' - 'git_commit_as_was' case does not work too
good. In fact equals the `commit' - 'git_commit_diff_to_parent', as this
is how git show works. This should be tried to improved in future.