fix: actually fix the selection index problem
Selection index does not necessarily mean the actual index in the
results table as while searching we get the filtered lists and the
selection index is according to that list.

Also, entry.index value does not change ever, so we have to loop through
the results to actually find the right index to remove.