Fix string.find() matching for only_cwd option
`string.find()` is defaulting to _pattern_ matching (rather than string
literal matching).  If you are using the `only_cwd` command in a
directory with a `-` (for example) the option fails to work.

This fix asks `string.find()` to interpret the arguments as literal
strings rather than patterns.

Reference: https://stackoverflow.com/a/15258515/181902