doc: clarify quickfixlist abbreviation in README (#1392)
This really killed me when I was looking at this plugin for the first time. I spent like 20 minutes googling how to get search results into the quickfixlist because ctrl-F for "quickfixlist" on the README yielded no results!

There's nothing wrong with using the abbreviation in some places, but it should appear at least once in the form of `abbreviation (abbr.)`. That is, spelling it out next to the abbreviation at least once shows the reader the abbreviation they can search for next to learn more.