fix: dont change highlights when using current_buffer_fuzzy_find
this happens if parsers are installed but highlight is set to false.
So parsing the buffer with treesitter will automatically apply them to
the buffer but because treesitter highlighting is disabled it will
change the buffer. Its stupid and i dont know why you want parser
installed but highlighting disabled but whatever.

We just dont calculate highlights for people who have highlight set
to false