Prevent an exception when err.message is nil
When trying to use go-to-definition using telescope's
LSP helper while Sumneko has not completed loading yet,
it returns an error response which does not have `message`
field. This error handler tries to log an error by accessing
`message` field which is a `nil`. That throws an exception in
Lua, and using `tostring()` prevents that.