fix(actions.qflist): make it work with vim quickfix commands (#1742)
use `r` mode to instead use `" "` mode. This makes Telescope quickfix list action send results
to a new list in the quickfix stack instead of overwriting the entries in the current list. 

Doing so enables `:chistory`, `:colder` and `:cnewer` functions to review results of 
previous Telescope queries.

The location list uses a similar API and is updated in this PR also
(ie this also enables :lhistory, :lolder and :lnewer).