feat: teach git_files to use the current repository
git_files only looks at the current repository, which is limiting
when editing files across several repositories.

A simple solution is to scope the "git ls-files" queries
to the current buffer's directory. This causes Telescope to
discover the git repository in the context of the current
file and return results relevant to that file.

Enable this behavior behind a new "use_per_file_cwd" so that
existing behavior is unchanged.

This is arguably better behavior in general and could be made the
default behavior, but is left as an opt-in feature.