Add use_buffer_path option to git_x pickers
This is helpfull when working on projects that use the repo tool.
In this type of projects, is a common practice that you start nvim
from the project root directory that contains several subdirs
corresponding to all the git repos that belongs to the project.

By leting telescope to find the git root directory based on the
current buffer path it's very easy to move across all the git
projects