fix(resolve): Check val not nil in resolve funcs (#2097)
We did not check `val ~= nil` in the resolve functions, so config like
`{ nil, max = 30 }` will throw a nil error. Also, if the config is `{
padding = _ }`, the logic relies on the function handling the padding is
iterated before the one handling min/max in the map, which is not always
guaranteed.

Fix the bug by adding nil check in the function handling min/max. Close