Allow `entry_maker` to return a flag to denote entries type
This is useful when one is creating a live_grep like picker that makes a
request to some url that returns a single json array with multiple
entries. From the perspective of the oneshot job, the json is on a
single line and has to be considered as a single entry, however this
might not be what we actually want. Instead, we might want to parse the
json and return *multiple* entries. For instance:
source line : "{someArray: [1, 2, 3]}" -> wanted entries : 1, 2, 3