Change lsp_document_symbols display
Before this patch, lsp document symbols would be displayed in two
columns:

1. A column with the 25 first characters in the symbol name.
2. A column with the symbol kind that could take up remaining
   characters.

That seems a little wrong-way-around and unnecessary cutting of symbol
names, which is why with this patch there are still two columns, but now
they are:

1. A column with the symbol kind (column size fitted to the longest name)
2. A column with the symbol name that takes up the remaining space.

It could have just stopped there, but it sort of lacked some icons. So
added is now configuration options for adding a symbol icon either in
front of or after the symbol's kind. These options are documented in
`config.lua`, but basically it is possible to:

* Have an icon or not (text will remain).
* Put the icon on either side of the text.
* Add space and or other characters around the icon.
* Choose which preset to use with `lspkind.nvim`
* Make substitutions on symbol kind, so if LSP reports kind `A` it can be
  translated to `B` in case `lspkind.nvim` has no icon for `A` but the
  one for `B` makes sense to use instead.
* Define what character(s) to fall back to in case an icon was not
  found.