Change the order of columns back to name, kind
Had a changeof heart regarding the order of these columns in the
lsp_document_symbol picker. The original order makes sense - it was just
the cutting off of symbol names that made it look wrong.

This patch changes the order back but makes it so the column with the
symbol name is wide enough to show the whole name.