feat(commands): sort autocomplete popup items
I noticed that neovim wasn't sorting the list of autocomplete items when pressing tab in the command line, e.g. `:Telescope <Tab>`.

In my very limited lua + nvim knowledge, there doesn't seem to be a way to create an autocommand that intercepts the completion list in the command popup.

So I decided to try my hand at brute-forcing the solution from telescope's side.

Here I sort the list of available commands for each of the 3 possible code paths.

Let me know if this is useful or if you have any other recommendations for sorting the options. It's also ok if this feature is not wanted. 👍