set default text to word in case of grep_string for picker to show the selected word
Before this PR, on running `grep_string` - the word under the cursor won't appear as the prompt text. Following is the use-case when it would be nice to have:

- User is searching for a word "entry"
- User decides to change something in the given word, quite common when searching for a test class. Example: searching for `GrepStringTest' word, user might decide to just remove the `Test` suffix to see if there is any `GrepString` word.

Currently (before this PR), the way to do this (please correct me if I'm wrong) is to select the word, paste in the prompt again and then remove the `Test` suffix (/type it themselves).

After this PR:

- User will see the word under the cursor as the text in the prompt.
- User can directly play around with the word, and doesn't need to do extra steps.

Why is this PR needed?

- IMO, I don't see a use-case where this would not be needed (ability to show the text, by default in the prompt).
- Reduces the efforts of the users who want to do a follow-up search.

How does this PR solve the issue?

- Sets `default_text` to the `word`. `default_text` is available as an argument while intialising a picker using `pickers::new`.