fix(git_status): keep picker open with no matches
Currently, `git_status` with a prompt that has zero matches closes the
picker.

Only automatically close the picker if working tree is completely clean.
If working tree is not clean, and a prompt is given but no matches are
found, keep the picker open. This behavior is consistent with other
pickers where lack of matches doesn't result in the picker closing.

closes #2815