fix(pickers): when no lnum given, don't set entry.lnum
Fixes couple of regressions from #2791 when no line number info is given
in the prompt (eg. `file.txt` rather than `file.txt:10`)
Both issues stemming from the fact that when no line number info is
given, the `entry.lnum` was set to `0`. `entry.lnum` is 1-index.
- Sending results to quickfix would send faulty results (#2864)
- Will no open the file on the correct (previous) line number
For this, setting the lnum to 1 is still undesirable since this
overwrites the lnum priority handling in the `select` action.