buffers picker: add select_current option
If select_current is true then the the buffers picker will
pre-select the current buffer (%) instead of the first one.
Useful if you want to use arrow keys to select a nearby buffer.
Best used in combination with selection_strategy = 'closest'
so that when you start typing the first option is selected.

Example usage:
```
  vim.keymap.set('n', '<M-b>', function()
    require('telescope.builtin').buffers {
      select_current = true,
      selection_strategy = 'closest',
    }
  end, { desc = 'Buffer list' })
```