fix(picker): prevent status text from hiding prompt (#2936)
There's a limitation with extmark's `right_align` virtual text and how
it handles side scrolling. This causes text entered to be hidden behind
the virtual text rather than begin the side scroll.

An issue is opened upstream https://github.com/neovim/neovim/issues/27540

But this could take a while before it's fixed as it's considered
`complexity:high`.

This PR adds a temporary workaround of just hiding the status text if the
prompt is long enough to interfere with it.