Fix find_references bug
There is a bug in list_or_jump implementation. The intended behavior is
to drop the item that corresponds to the current location when the
option is to not include the current location in the list.

However, a bug in implementation assumes that the locations and items
list are in the same order, which is not the case, as location_to_items
implementation sorts the items. So the result is that often the wrong
item got dropped. When there are more than 1 references, the behavior
would be potentially missing reference. When there is 1 reference, it
could be potentially not jumping at all when the reference instead of
the definiion got dropped.

This fix takes a table of items sorts the locations in the same order as
items. It has O(N) complexity where N is the size of the items table.