fix(previewer): `preview=false` option inheritance (#3327)
According to `:h telescope.defaults.preview`, we should be able to do
`preview=false` to disable the previewer for pickers. With how picker
options resolve inheritance, something like this should work.

```lua
require('telescope').setup {
  defaults = { preview = true },
  pickers = { find_files = { preview = false } }
}
```
Here, find_files should not show a previewer but it previously would.

Corrects this by checking if `opts.preview` is explicitly set to
`false`. In which case, we won't show the previewer.

Closes #3325